Changes to local song database to properly reflect data changes..
Once the scan for audio files is complete, we now perform a bunch of checks to see what the nature of changes between the existing database data, and the newly scanned data is.

Then, we perform updates, insertions and deletions as required. This means that song ids are persisted even if the contents of the song has changed, and we can detect changes to the last modified date of a song, and propagate any changes to the database.

There is currently a limitation where, if the name of an album-artist has changed, or the name or album-artist name of an album has changed, the album is treated as new, and the old album is deleted. This can be resolved by implementing a contents equality check in the diff function in LocalSongRepository.updateAlbumArtistDatabase() and LocalSongRepository.updateAlbumDatabase(), which can check whether the albums/songs have changed, and if not, treat the change as an update rather than a delete and insert.