When calling pause, ensure the 'next media player' is also paused..
It may be about to commence playback, but it's not the current media player. In this case, calling pause() would result in continued playback.

Also, call `onPlaybackComplete()` after transitioning to the next player & updating the queue, to ensure any listeners updating themselves in response to onPlaybackComplete() are using up-to-date information.