Fixed a playback fragment layout issue..
The artwork item was rendering slightly taller than the available space, which made it vertically scrollable.

Turns out there's no need to embed the RV inside a NestedScrollView, as RV has its own 'nestedScrollingEnabled()' property, which we can disable to allow smooth interaction with the Bottom Sheet Behavior. Moving RV out of NSV and adjusting the list item to have 0dp for both width and height seems to resolve the issue.