LocalSongRepository refactoring..
The LocalSongRepository database is now common, with the actual song importing delegated to an implementation of SongProvider. There will only be one Shuttle song database, but there may be many different song providers. MockSongRepository has been refactored to MockSongProvider, and an implementation of MediaStore SongProvider has been created (but is not in use).

Artist, album and song names are now considered equal regardless of case.

Fixed an issue where a database insert conflict when inserting an artist/album could cause the associated songs to fail to insert, as the song's artist/album ids would resolve to '-1' (causing a foreign constraint issue).