Fixed an issue where shuffling didn't update the next-track-to-play..
We were loading a list of songs, which would turn shuffle off. Then we'd turn shuffle on when the load was complete, which would reshuffle the queue, but wouldn't update the playback manager's next-track.

Instead, we now call the overloaded load() method, passing a pre-shuffled list of songs. Since we're passing a shuffle list, we don't turn shuffle off, so we don't have to wait for the load to complete before turning shuffle on again.

Additionally, there's now a boolean overload on queueManager.setShuffleMode() called reshuffle. If this is false, the base (unshuffled) queue is not shuffled. Useful for when an already shuffled list has been passed to the queue manager, to avoid a redundant reshuffling.