Added setting option for theming..
Fixed a bunch of navigation issues

The Bottom Drawer is now a DialogFragment, and is navigated to using the NavController

Fixed an issue where nested navigation fragment was added to the wrong fragment manager, due to living inside the xml of the custom view's merge layout which was inflated in init() of the custom view. Long story.