Don't release playback manager when service is destroyed..
The playback manager is used outside of the scope of the service, any time music isn't playing and position/duration/playstate are queried, this needs to be available. Since PlaybackManager is singleton, and it's required for the lifespan of the app, we can't know when a user is completely finished with playback, and it would be a waste of resources to have to spin it up any time playback is queried.