Migrated to Coroutines..
This change replaces RxJava with coroutines.

There's also a change to the MediaDatabase, and the way songs are imported. The album and albumArtist tables have been dropped, and were now using Sqlite Views instead. Since albums and albumArtists are just derived from songs anyway, this saves us having to maintain their tables. Allows for easy updating of song data in the database.