Revert change to QueueBinder playback watcher..
It seems inefficient (and possibly causing lag throughout the app) when updating QueueBinder items constantly in response to progress updates, using the adapter.notify() function. So we revert to using a PlaybackWatcherCallback directly on the QueueBinder.ViewHolder, and ensure we set the recyclerview and its adapter to null when the fragment is destroyed, to prevent leaks.