MediaSession improvements.. (fixed bluetooth issue)
The MediaSession no longer spams updates when the playback progress changes.

Since the onQueueItemChanged() callback occurs before the ExoPlayer Playback has loaded the next song, updating MediaSession PlaybackState at this time, with the current Playback progress, resulted in the wrong progress being posted for the new song.

Now, when the queue item changes, if we know the previous 'current' queue item has changed (current song is different in response to a queue item change), we manually set the playback progress to 0.

This means we no longer have to update the MediaSession PlaybackState every time progress changes (which is every 100ms). This in turn improves performance and resolves and issue where spamming the AVRCP receiver caused it to fail to display the correct metadata.