Playback service startForeground/notification changes..
startForeground() is now called whenever any intent is passed to onStartCommand(). This ensures Android doesn't ANR the service.

If onStartCommand() is called with an intent, and the queue has not been loaded, we show a 'loading' notification. If the queue has been loaded, but it is empty, we show a 'queue empty' notification. In case the command will eventually load the queue, we keep displaying this notification - non-dismissable - until either the notification is replaced with a more useful one, or 10 seconds has passed.

Widget improvements:

- Fixed an issue where widget information was cleared when queue loads
- Show S2 launcher icon as placeholder when queue is empty
- Hide next/prev/play/pause buttons when queue is empty