Fixed a crash on API 31 (Foreground Service)
Google introduced a new type of crash, where the system kills your app if you try to start a foreground service without any input from the user.

This happens when audio focus is regained after transient focus loss (like a phone call), and we try to start the foreground service again.

The solution is to not stop the foreground service when pausing due to transient focus loss